+++ src/coordinator/users.spec.js
@@ -851,7 +851,7 @@ describe('Coordinator/Users', function() {
     });
   });
 
-  describe('syncWithAuth0', function() {
+  describe('sync', function() {
     context('when all required parameters are present', function() {
       let user;
       let promise;
@@ -860,10 +860,10 @@ describe('Coordinator/Users', function() {
         user = fixture.build('contxtUser');
 
         const users = new Users(baseSdk, baseRequest, expectedHost);
-        promise = users.syncWithAuth0(user.id);
+        promise = users.sync(user.id);
       });
 
-      it('sends a request to sync auth0 access', function() {
+      it('sends a request to sync user permissions', function() {
         expect(baseRequest.get).to.be.calledWith(
           `${expectedHost}/users/${user.id}/sync`
         );
@@ -877,10 +877,10 @@ describe('Coordinator/Users', function() {
     context('when the user ID is not provided', function() {
       it('throws an error', function() {
         const users = new Users(baseSdk, baseRequest, expectedHost);
-        const promise = users.syncWithAuth0(null);
+        const promise = users.sync(null);
 
         return expect(promise).to.be.rejectedWith(
-          'A user ID is required for syncing with auth0'
+          'A user ID is required for syncing user permissions'
         );
       });
     });